requirements.txt
anthropic>=0.25.0
notion-client>=2.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
schedule>=1.2.0
python-dotenv>=1.0.0
